refactor: 見積申請の整形ヘルパを一覧・詳細で共有化
formatDateTime / formatYen / SUBMISSION_TYPE_LABELS を一覧固有の
_components/labels.ts から一覧・詳細の共通親 _shared/labels.ts へ移設。
詳細画面(#574)でも同じ提出区分ラベル・金額/日時整形を使うため。
機能固有ゆえ global(_lib)へは昇格しない。
